backend/requirements.txt
aiofiles==24.1.0
aiosmtplib==3.0.1
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
black==23.12.1
boto3==1.34.129
botocore==1.34.162
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==42.0.8
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.1.1
fastapi==0.104.1
flake8==6.1.0
h11==0.14.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.4
jmespath==1.0.1
jq==1.10.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mypy==1.8.0
mypy_extensions==1.1.0
numpy==1.24.4
oauthlib==3.3.1
packaging==25.0
pandas==2.1.4
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.1.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.23
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.5.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.10.4
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2024.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
slowapi==0.1.9
httpx==0.25.0
bleach==6.0.0
